+++ refract/__init__.py
@@ -1,10 +1,10 @@
 from refract.elements import (Element, Metadata, String, Number, Boolean, Null,
                               Array, Member, KeyValuePair, Object)
-from refract.namespace import Namespace
+from refract.registry import Registry
 from refract.refraction import refract
 
 
 __all__ = (
     'Element', 'Metadata', 'String', 'Number', 'Boolean', 'Null',
-    'Array', 'Member', 'KeyValuePair', 'Object', 'Namespace', 'refract'
+    'Array', 'Member', 'KeyValuePair', 'Object', 'Registry', 'refract'
 )